requirements.txt
langchain
openai
langchain-community
python-dotenv
streamlit
pypdf
numpy
langchain-openai